홈타입 수정, 헬스장 운동 추천 AI와의 API통신에 필요한 로직 구성, GymReport에 Domain에  Today타입 추가

+++ src/main/java/fitrecommend/fitquest/domain/Exercise.java
@@ -32,10 +32,10 @@ public class Exercise {
 
     private Complete complete;
 
-    private Integer totalKcal;
+    private Long totalKcal;
 
-    public int getTotalKcal(){
-        int kcal = 0;
+    public Long getTotalKcal(){
+        Long kcal = 0L;
         for(Set set : sets){
             kcal +=  set.getRep() * gym.getKcal();
         }
